NNUE (#30)
### 8+0.08 (STC)
```
Score of svart3 vs svart2 after 512 rounds: 855 - 41 - 128 (0.90)
Ptnml:        WW     WD  DD/WL     LD     LL
Distr:       357    110     35     10      0
Stats:  W: 83.5%   L: 4.0%   D: 12.5%   TF: 0
Elo difference: 376.85 +/- 28.65
```

### 40+0.4 (LTC)
```
Score of svart3 vs svart2 after 700 rounds: 1150 - 40 - 210 (0.90)
Ptnml:        WW     WD  DD/WL     LD     LL
Distr:       474    172     45      8      1
Stats:  W: 82.1%   L: 2.9%   D: 15.0%   TF: 0
Elo difference: 374.91 +/- 23.12
```

+++ src/engine/history.rs
@@ -1,6 +1,6 @@
 use cozy_chess::{Board, Move};
 
-pub const MAX_HISTORY: i16 = i16::MAX;
+pub const MAX_HISTORY: i32 = i16::MAX as i32;
 
 pub struct History {
     pub table: [[[i32; 64]; 64]; 2],
@@ -22,14 +22,16 @@ impl History {
         self.table[color][from][to]
     }
 
-    pub fn update_table<const POSITIVE: bool>(&mut self, board: &Board, mv: Move, depth: i16) {
-        let delta = (16 * (depth * depth)).min(1200) as i32;
+    pub fn update_table<const POSITIVE: bool>(&mut self, board: &Board, mv: Move, depth: i32) {
+        let delta = (16 * (depth * depth)).min(1200);
         let bonus = if POSITIVE { delta } else { -delta };
+
         self.update_score(board, mv, bonus);
     }
 
     pub fn update_score(&mut self, board: &Board, mv: Move, bonus: i32) {
-        let scaled_bonus = bonus - self.get_score(board, mv) * bonus.abs() / MAX_HISTORY as i32;
+        let scaled_bonus = bonus - self.get_score(board, mv) * bonus.abs() / MAX_HISTORY;
+
         let color = board.side_to_move() as usize;
         let from = mv.from as usize;
         let to = mv.to as usize;
